requirements.txt
python-multipart
ultralytics
opencv-python
pillow
matplotlib
pyyaml
pytest
PyMuPDF
fastapi
uvicorn
